refine doc and code
