+++ onediff_sd_webui_extensions/README.md
@@ -98,13 +98,13 @@ If you possess a OneDiff Enterprise license key, you can access instructions on 
 
 Note: The quantified model doesn't currently support LoRA loading, and related features are under emergency development.
 
-## API
+## Use OneDiff by API
 
 To use the OneDiff-based sd-webui API, you only need to add a `"script": "onediff_diffusion_model"` field to the request to speed up model inference.
 
 For OneDiff Enterprise, add `"script_args" : [{"0": True}]` to use the quantization feature.
 
-Check file `extensions/onediff_sd_webui_extensions/api.py` for more details.
+Check file `./onediff_sd_webui_extensions/api_examples/txt2img.py` for more details.
 
 Run the commands below to use Sable Diffusion WebUI with OneDiff extensions.
 
@@ -116,9 +116,7 @@ python3 webui.py --api
 python3 extensions/onediff_sd_webui_extensions/api.py
 ```
 
-Then you can get the images returned by sd-webui client at `api_out/txt2img/`.
-
-Note: OneDiff sd-webui only supports txt2img now, img2img and ControlNet is under urgent development.
+Then you can get the images returned by sd-webui client at `./api_out/txt2img/`.
 
 ## Contact
